+++ MortalFighter/Sensor.h
@@ -2,7 +2,7 @@
 class Sensor
 {
 public:
-	Sensor(std::pair<float, float> posicion, int alto, int ancho, bool esHitBox, std::string unestadoStr);
+	Sensor(std::pair<int, int> posicion, int alto, int ancho, bool esHitBox, std::string unestadoStr);
 	
 	void setPosicion(std::pair<float, float> posicion);
 	void setAlto(int alto);
